feat(CalloutMenu): add menu to TeaserFeed element
